i do le comit with  shit from futur

+++ Proj/RType/RType/ServerCommunication.cpp
@@ -183,9 +183,9 @@ template<class T>
 void ServerCommunication<T>::UDPscreenState(Packet& packet, unsigned int score, std::list<Element*>& elements) // elements pour idSprite et CoordSprite
 {
 	char opcode = Opcodes::screenState;
-	unsigned short datasize = htons(sizeof(unsigned int) + (static_cast<unsigned short>(elements.size()) * (sizeof(u_long) + sizeof(t_coord))));
+	unsigned short datasize = htons(sizeof(unsigned int) + (static_cast<unsigned short>(elements.size()) * (sizeof(unsigned int) + sizeof(t_coord))));
 	unsigned int score_to_send = htonl(score);
-	u_long id = 24;
+	unsigned int id = 24;
 	t_coord coord;
 	std::list<Element*>::const_iterator ite = elements.begin();
 
@@ -199,7 +199,7 @@ void ServerCommunication<T>::UDPscreenState(Packet& packet, unsigned int score, 
 		id = htonl((*ite)->getSprite().front());
 		coord._posX = htons((*ite)->getPos()._posX);
 		coord._posY = htons((*ite)->getPos()._posY);
-		packet.write(reinterpret_cast<char*>(&id), sizeof(u_long));
+		packet.write(reinterpret_cast<char*>(&id), sizeof(unsigned int));
 		packet.write(reinterpret_cast<char*>(&coord), sizeof(t_coord));
 		++ite;
 	}
